make template loader look for PYME templates first. Make plugin template access require referencing the template by plugin name. Allow plugins to specify a template directory, rather than individual templates. Remove python code from the template directory. TODO - make accompanying changes to PYME.config
